refactor: move statement components outside of AfRowLayoutDesktop
